pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "Modular RAG platform backend services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "structlog",
    "python-dotenv",
    "sqlalchemy>=2.0",
    "alembic",
    "psycopg[binary]",
    "qdrant-client",
    "tenacity",
    "typer[all]",
    "python-magic",
    "tiktoken",
    "pypdf",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "mypy",
    "pytest-httpx",
    "pytest-mock",
]

[project.scripts]
rag = "backend.cli:app"

[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend"]

[tool.uvicorn]
app = "backend.app.main:app"
host = "0.0.0.0"
port = 8000
reload = true
factory = false

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
strict = false
